requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
Pillow>=10.2.0
pyotp==2.9.0
qrcode==7.4.2
twilio==8.10.0
sendgrid==6.10.0
django-filter==23.4
drf-spectacular==0.26.5
psycopg[binary]==3.2.9
